+++ modules/spring-core/src/main/kotlin/com/explyt/spring/core/SpringCoreClasses.kt
@@ -140,6 +140,8 @@ object SpringCoreClasses {
 
     const val MESSAGE_MAPPING = "org.springframework.messaging.handler.annotation.MessageMapping"
 
+    const val HIBERNATE_CFG_AVAILABLE_SETTING = "org.hibernate.cfg.AvailableSettings"
+
     val ANNOTATIONS_WITH_FILE_REFERENCES_TO_PROPERTIES = setOf(PROPERTY_SOURCE, TEST_PROPERTY_SOURCE)
     val ANNOTATIONS_WITH_FILE_REFERENCES_TO_XML = setOf(CONTEXT_CONFIGURATION, IMPORT_RESOURCE)
     val ANNOTATIONS_WITH_FILE_REFERENCES_TO_SQL = setOf(CONTEXT_SQL)