+++ modules/spring-core/src/main/kotlin/com/explyt/spring/core/properties/references/ValueHintReference.kt
@@ -21,7 +21,11 @@ import com.explyt.spring.core.JavaCoreClasses
 import com.explyt.spring.core.SpringCoreClasses
 import com.explyt.spring.core.SpringProperties
 import com.explyt.spring.core.SpringProperties.VALUES
-import com.explyt.spring.core.completion.properties.*
+import com.explyt.spring.core.completion.properties.PropertyHint
+import com.explyt.spring.core.completion.properties.ProviderHint
+import com.explyt.spring.core.completion.properties.ProviderParameters
+import com.explyt.spring.core.completion.properties.SpringConfigurationPropertiesSearch
+import com.explyt.spring.core.completion.renderer.PropertyValueRenderer
 import com.explyt.spring.core.properties.ClassReferencePropertyRenderer
 import com.explyt.spring.core.service.SpringSearchServiceFacade
 import com.explyt.spring.core.util.PropertyUtil